+++ Runtime/Scripts/Utilities/GUILayoutUtility.cs
@@ -4,7 +4,7 @@ namespace Unity.FilmInternalUtilities {
 
 internal static class GUILayoutUtility {
 
-    static void SetWithLastRect(ref Rect rect) {
+    internal static void SetWithLastRect(ref Rect rect) {
         Rect r = UnityEngine.GUILayoutUtility.GetLastRect();
         if (Event.current.type == EventType.Repaint) {
             rect = r;
@@ -13,7 +13,7 @@ internal static class GUILayoutUtility {
     
 //----------------------------------------------------------------------------------------------------------------------    
 
-    static void ReserveRect(ref Rect rect, GUIContent content, GUIStyle guiStyle, params GUILayoutOption[] options) {
+    internal static void ReserveRect(ref Rect rect, GUIContent content, GUIStyle guiStyle, params GUILayoutOption[] options) {
         Rect r = UnityEngine.GUILayoutUtility.GetRect(content, guiStyle, options);
         if (Event.current.type == EventType.Repaint) {
             rect = r;