+++ Tests/Runtime/Scripts/Timeline/DummyTimelineTrack.cs
@@ -8,6 +8,10 @@ namespace Unity.FilmInternalUtilities.Tests {
 [TrackClipType(typeof(DummyTimelinePlayableAsset))]
 [Serializable]
 internal class DummyTimelineTrack : BaseExtendedClipTrack<DummyTimelineClipData> {
+
+    protected override void OnEnableInternalV() {
+//        Debug.Log("DummyTimelineTrack::OnEnableInternalV()");
+    }
     
     protected override Playable CreateTrackMixerInternal(PlayableGraph graph, GameObject go, int inputCount) {
         return Playable.Create(graph, inputCount);